Added setCatalog to "public"
Use setCatalog on the connection to set the catalog to be used in subsequent calls. This makes it so that it can in theory only "see" the tables that it is set to. The problem with it not working for PostgreSQL possibly could be related to the table names being uppercase and JDBC thinking it knows better.

**Note**. Oracle always uppercases table names, PostgreSQL always uses lowercase.